# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/scrapers": {
      "get": {
        "operationId": "ScrapersController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit",
            "schema": {
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of scrapers with pagination",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScrapersListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Scrapers"
        ]
      },
      "post": {
        "operationId": "ScrapersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScraperCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created scraper",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScraperDto"
                }
              }
            }
          }
        },
        "tags": [
          "Scrapers"
        ]
      }
    },
    "/api/scrapers/{id}": {
      "get": {
        "operationId": "ScrapersController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scraper by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScraperDto"
                }
              }
            }
          }
        },
        "tags": [
          "Scrapers"
        ]
      },
      "patch": {
        "operationId": "ScrapersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScraperUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated scraper",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScraperDto"
                }
              }
            }
          }
        },
        "tags": [
          "Scrapers"
        ]
      },
      "delete": {
        "operationId": "ScrapersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete result",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Scrapers"
        ]
      }
    },
    "/api/scrapers/{id}/run": {
      "post": {
        "operationId": "ScrapersController_run",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Enqueue scraping job for the scraper by id",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "jobId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Scrapers"
        ]
      }
    }
  },
  "info": {
    "title": "Trade API",
    "description": "Auto-generated OpenAPI documentation for Trade backend",
    "version": "1.0.0",
    "contact": {
      "name": "API",
      "url": "",
      "email": ""
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "schemas": {
...
      "AugmentedZodDto": {
        "type": "object",
        "properties": {
          "API": {
            "type": "object",
            "properties": {
              "url": {
                "type": "string",
                "format": "uri"
              }
            },
            "selfRequired": true
          },
          "HTML": {
            "type": "object",
            "properties": {
              "url": {
                "type": "string",
                "format": "uri"
              },
              "selectors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "selector": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "name",
                    "selector"
                  ]
                }
              },
              "headers": {
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {
                  "type": "string"
                }
              },
              "timeoutMs": {
                "type": "number",
                "exclusiveMinimum": 0
              },
              "pagination": {
                "type": "object",
                "properties": {
                  "nextSelector": {
                    "type": "string",
                    "minLength": 1
                  },
                  "nextUrlTemplate": {
                    "type": "string",
                    "minLength": 1
                  },
                  "pageParam": {
                    "type": "string",
                    "minLength": 1
                  },
                  "startPage": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 9007199254740991
                  },
                  "maxPages": {
                    "type": "integer",
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991
                  }
                }
              },
              "document": {
                "type": "object",
                "properties": {
                  "linkSelector": {
                    "type": "string"
                  },
                  "linkAttr": {
                    "type": "string"
                  },
                  "titleSelector": {
                    "type": "string"
                  },
                  "contentSelector": {
                    "type": "string"
                  },
                  "baseUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "maxDocsPerPage": {
                    "type": "integer",
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991
                  }
                },
                "required": [
                  "linkSelector"
                ]
              }
            },
            "selfRequired": true
          }
        },
        "required": [
          "API",
          "HTML"
        ]
      },
      "ScraperDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "API",
              "HTML"
            ]
          },
          "config": {
            "$ref": "#/components/schemas/AugmentedZodDto"
          },
          "postProcessors": {
            "description": "Post-processors list",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "config"
        ]
      },
      "ScrapersListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScraperDto"
            }
          },
          "total": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "ScraperCreateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Scraper name"
          },
          "type": {
            "type": "string",
            "description": "Scraper type",
            "enum": [
              "API",
              "HTML"
            ]
          },
          "config": {
            "description": "Scraper config",
            "allOf": [
              {
                "$ref": "#/components/schemas/AugmentedZodDto"
              }
            ]
          },
          "postProcessors": {
            "description": "Post-processors list",
            "example": [
              {
                "type": "TRIM_WHITESPACE",
                "config": {
                  "trimEachLine": true
                }
              }
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "type",
          "config"
        ]
      },
      "ScraperUpdateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Scraper name"
          },
          "type": {
            "type": "string",
            "description": "Scraper type",
            "enum": [
              "API",
              "HTML"
            ]
          },
          "config": {
            "description": "Scraper config",
            "allOf": [
              {
                "$ref": "#/components/schemas/AugmentedZodDto"
              }
            ]
          },
          "postProcessors": {
            "description": "Post-processors list",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}